+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Dependencies
 - Bumps `org.apache.httpcomponents.core5:httpcore5-h2` from 5.2.3 to 5.2.4
 - Bumps `org.owasp.dependencycheck` from 9.0.9 to 9.1.0
-- Bumps `io.github.classgraph:classgraph` from 4.8.165 to 4.8.170
+- Bumps `io.github.classgraph:classgraph` from 4.8.165 to 4.8.172
 - Bumps `org.eclipse.parsson:parsson` from 1.1.5 to 1.1.6
 
 This section is for maintaining a changelog for all breaking changes for the client that cannot be released in the 2.x line. All other non-breaking changes should be added to [Unreleased 2.x] section.