+++ CHANGELOG.md
@@ -26,7 +26,8 @@ This section is for maintaining a changelog for all breaking changes for the cli
 - Delete shape property ([#884](https://github.com/opensearch-project/opensearch-java/pull/885))
 
 ### Fixed
-- Fix version and build ([#254](https://github.com/opensearch-project/opensearch-java/pull/254))
+-  Fix integer overflow for variables in indices stats response ([#959](https://github.com/opensearch-project/opensearch-java/pull/96))
+
 
 ### Security